move CRServiceWorker to own file
